+++ diachron-archive/archive-data-mapping/src/main/java/eu/fp7/diachron/archive/mapping/MappingUtils.java
@@ -1,4 +1,4 @@
-package eu.fp7.diachron.mapper;
+package eu.fp7.diachron.archive.mapping;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
